requirements.txt
environs==14.5.0
fastapi==0.128.8
psycopg2-binary==2.9.11
asyncpg==0.31.0
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.46
starlette==0.52.1
uvicorn==0.40.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
bcrypt==4.0.1
email-validator==2.3.0
python-multipart==0.0.22
aiofiles==25.1.0

pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
pytest-mock==3.15.1
faker==40.5.1